# optimize ComputeDecibel in fsmn-vad model by using numpy